DESCRIPTION - sorted files and functions - removing quotes WIP
