+++ src/parsing/my_split.c
@@ -6,7 +6,7 @@
 /*   By: jebucoy <jebucoy@student.42abudhabi.ae>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/05 18:03:07 by jebucoy           #+#    #+#             */
-/*   Updated: 2023/09/05 18:17:30 by jebucoy          ###   ########.fr       */
+/*   Updated: 2023/09/22 18:56:22 by jebucoy          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,23 +32,6 @@ static size_t	str_count(char const *s, char c)
 		i++;
 	}
 	return (count);
-}
-
-// qflag = 0
-// if single quote (') is encounted, set it to one
-// if double quote (") set it to 2
-// if a closing quote of the specific is encountered, reset to 0
-int	is_quotes_closed(int qflag, char q)
-{
-	if (q == SINGLE && qflag == 0)
-		return (1);
-	else if (q == DOUBLE && qflag == 0)
-		return (2);
-	else if (q == SINGLE && qflag == 1)
-		return (0);
-	else if (q == DOUBLE && qflag == 2)
-		return (0);
-	return (qflag);
 }
 
 // counts the characters in s[i] for malloc
