+++ src/parsing/utils.c
@@ -6,7 +6,7 @@
 /*   By: jebucoy <jebucoy@student.42abudhabi.ae>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/13 19:46:27 by jebucoy           #+#    #+#             */
-/*   Updated: 2023/09/09 19:03:54 by jebucoy          ###   ########.fr       */
+/*   Updated: 2023/09/22 19:05:50 by jebucoy          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
